u-boot: fix sha1 header include
The include directory is added to the list of header files to be
searched with the -idirafter option. This option marks the directory as
a system include directory, which seems to break the quoted include of the
sha1 header file.

Including the directory with -I breaks things in other ways, so let's
just include it with its relative path.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>